fixed typing

+++ sklearn/decomposition/nmf.py
@@ -475,8 +475,8 @@ def _fit_coordinate_descent(X, W, H, tol=1e-4, max_iter=200, l1_reg_W=0,
     for n_iter in range(max_iter):
         violation = 0.
 
-        w_free_cols = np.ones((W.shape[1],))
-        ht_free_cols = np.ones((H.shape[0],))
+        w_free_cols = np.ones((W.shape[1],)).astype('int32')
+        ht_free_cols = np.ones((H.shape[0],)).astype('int32')
 
         # Update W
         violation += _update_coordinate_descent(X, W, Ht, l1_reg_W,
